MAC validation,  DeviceFilter.__repr__
this MAC validation is not standard compliant, because formats like aa-bb:ccdd.ee:ff are also supported as long as it parses to a valid hex string, but should be "good enough" for the moment. I measured the performance of regex vs this validation and the regex takes slightly more time in the OK case, which i assumed is the more relevant case.
added a __repr__ method to the DeviceFilter for more meaningful debugging output.